dashboard/requirements.txt
dash
dash-bootstrap-components
pandas
python-dotenv
plotly
scipy
scikit-posthocs
statsmodels
gunicorn
kaleido
